moving to download.go, and removing check on code download
